move all ModelResource objects from admin to export.resources
