+++ parlacards/admin/number_of_spoken_words.py
@@ -1,24 +1,11 @@
 from django.contrib import admin
 
-from import_export.resources import ModelResource
 from import_export.admin import ExportMixin
-from import_export.fields import Field
+
+from export.resources import PersonNumberOfSpokenWordsResource
 
 from parlacards.models import PersonNumberOfSpokenWords
-
 from parlacards.admin.common import LatestScoresAdmin
-
-
-class PersonNumberOfSpokenWordsResource(ModelResource):
-    name = Field()
-
-    class Meta:
-        model = PersonNumberOfSpokenWords
-        fields = ('name', 'value', 'timestamp',)
-        export_order = ('name', 'value', 'timestamp',)
-    
-    def dehydrate_name(self, score):
-        return score.person.name
 
 
 # TODO this could be abstracted away into a sort of current score admin
