+++ parladata/admin/person.py
@@ -5,9 +5,9 @@ from django.utils.safestring import mark_safe
 from django.urls import reverse
 from django.conf import settings
 
-from import_export import resources
 from import_export.admin import ExportMixin
-from import_export.fields import Field
+
+from export.resources import MPResource
 
 from parladata.models import Person, PersonMembership
 from parladata.models.versionable_properties import *
@@ -85,34 +85,7 @@ class ParliamentMember(Person):
         proxy = True
 
 
-class MPResource(resources.ModelResource):
-    name = Field()
-    age = Field()
-    education_level = Field()
-    preferred_pronoun = Field()
-    number_of_mandates = Field()
 
-    class Meta:
-        model = Person
-        fields = ('id', 'name', 'date_of_birth', 'age', 'education_level', 'preferred_pronoun', 'number_of_mandates',)
-        export_order = ('id', 'name', 'date_of_birth', 'age', 'education_level', 'preferred_pronoun', 'number_of_mandates',)
-    
-    def dehydrate_name(self, person):
-        return person.name
-
-    def dehydrate_age(self, person):
-        if person.date_of_birth:
-            return int((datetime.now().date() - person.date_of_birth).days / 365.2425)
-        return None
-    
-    def dehydrate_education_level(self, person):
-        return person.education_level
-    
-    def dehydrate_preferred_pronoun(self, person):
-        return person.preferred_pronoun
-    
-    def dehydrate_number_of_mandates(self, person):
-        return person.number_of_mandates
 
 class MPAdmin(ExportMixin, PersonAdmin):
     resource_class = MPResource
